portfolio cleanup
